module_6/src/web/requirements.txt
# Flask Web Framework
Flask==3.1.3
# Pin Jinja2 to fix XSS/Template Injection vulnerabilities
Jinja2==3.1.6

# PostgreSQL Database Adapter
psycopg2-binary==2.9.9

# Environment variable management
python-dotenv==1.2.1

# RabbitMQ client
pika==1.3.2
